+++ hazelcast-raft-server/src/main/java/com/hazelcast/raft/service/atomiclong/operation/CompareAndSetOperation.java
@@ -26,7 +26,7 @@ public class CompareAndSetOperation extends AbstractAtomicLongOperation {
     }
 
     @Override
-    public Object doRun(int commitIndex) {
+    public Object doRun(long commitIndex) {
         RaftAtomicLong atomic = getAtomicLong();
         return atomic.compareAndSet(currentValue, newValue, commitIndex);
     }
